Add selectedAccount and currentAccount
