+++ FRW/Modules/Profile/WalletSetting/SecureEnclavePrivateKeyView.swift
@@ -92,7 +92,7 @@ struct SecureEnclavePrivateKeyView: RouteableView {
                     }
                 }
 
-                if let key = WalletManager.shared.currenFlowAccount?.fullWeightKey {
+                if let key = WalletManager.shared.currentMainAccount?.fullWeightKey {
                     HStack {
                         HStack(spacing: 16) {
                             Divider()
