+++ FRW/Services/Manager/PushHandler.swift
@@ -22,7 +22,7 @@ class PushHandler: NSObject, ObservableObject {
 
         refreshPushStatus()
 
-        WalletManager.shared.$walletInfo
+        WalletManager.shared.$currentMainAccount
             .dropFirst()
             .debounce(for: .seconds(1), scheduler: DispatchQueue.main)
             .receive(on: DispatchQueue.main)
