+++ FRW/Services/Manager/Wallet/WalletManager+Storage.swift
@@ -1,0 +1,8 @@
+//
+//  WalletManager+Storage.swift
+//  FRW
+//
+//  Created by Hao Fu on 31/3/2025.
+//
+
+import Foundation
