+++ FRW/Tools/Helper+Extensions.swift
@@ -19,3 +19,19 @@ extension Optional<URL> {
         self?.absoluteString.isEmpty == false
     }
 }
+
+extension Optional: @retroactive RawRepresentable where Wrapped: Codable {
+    public var rawValue: String {
+        guard let data = try? JSONEncoder().encode(self) else {
+            return "{}"
+        }
+        return String(decoding: data, as: UTF8.self)
+    }
+
+    public init?(rawValue: String) {
+        guard let value = try? JSONDecoder().decode(Self.self, from: Data(rawValue.utf8)) else {
+            return nil
+        }
+        self = value
+    }
+}
